Optionally include kdebuild-1 documentation.
kdebuild-1 is a new EAPI requested by the Gentoo KDE project for use in their
overlay. Having documentation of this format is useful for package manager
writers and kdebuild-1 ebuild authors, but we may or may not want to include it
in the PMS copy we'll submit to the Council. So we have options in pms.tex to
enable or disable inclusion -- edit the value of the ENABLE-KDEBUILD boolean as
appropriate (for now it's on by default). We also have an option to show what
PMS looks like both with and without ENABLE-KDEBUILD, by showing both sides of
conditionals in different colours. This is the ENABLE-ALL-OPTIONS boolean,
which is also enabled by default.

Squashed commit of the following:

commit 137e55e3a2712e5b1f9cbb5d168634ca68b2dbe9
Author: Ciaran McCreesh <ciaran.mccreesh@googlemail.com>
Date:   Tue Mar 18 07:24:14 2008 +0000

    Wording tweak

commit dcd912651f6d0bda15631ec3f16a306740decebc
Author: Ciaran McCreesh <ciaran.mccreesh@googlemail.com>
Date:   Tue Mar 18 06:56:25 2008 +0000

    Rewrite the "no magic mkdir" rules.

    The ban on magic mkdir voodoo gets really messy definition-wise if it's put on
    'into', since we don't want it to apply where the package manager is explicit.
    Reword it so that it just bans on explicit utilities.

commit e39c9d23614a37c3cd6f7e336598f27712cd4b0a
Author: Ciaran McCreesh <ciaran.mccreesh@googlemail.com>
Date:   Tue Mar 18 06:22:12 2008 +0000

    Optionally include kdebuild-1 documentation.