Rework tables for EAPI 3.
Update all feature tables for EAPI 3 as a copy of EAPI 2, and introduce
an empty change list for EAPI 3. Things differing in EAPI 3 will be
changed in subsequent commits.

A few tables used the order 0, 1, 2, kdebuild-1 rather than 0, 1,
kdebuild-1, 2. These are now fixed for consistency.

Switch the features table to landscape, to make it fit better.