Move bibliography before the appendix.
Style manuals have different opinions about the position of the
bibliography: The Chicago Manual of Style says it should be placed
after the appendix; APA and Harvard say it should be placed before.
Move it from its current position (between appendix E and F) to before
the appendix, because it would look strange after the EAPI cheat
sheet.

Also make the bibliography appear in the table of contents, by means
of the "tocbibind" package.