EAPI 6: Ensure sane settings for LC_CTYPE and LC_COLLATE.
Some locales (e.g. tr_TR) change the upper/lowercase mapping for ASCII
characters, which breaks the tr(1) command and the case-modifying word
expansions of Bash version 4.

Therefore, ensure that LC_CTYPE and LC_COLLATE behave in the same way
as the POSIX locale, for characters in the ASCII range.