Rearrange front matter.
Move sections Reporting Issues and Acknowledgments to the Introduction
chapter (and number them). Copyright and licence information goes to
the verso of the title page.

For the printed version this follows common practice of books, while
keeping the text in front of the table of contents reasonably short
for the HTML version.