Restore proper scope for PORTDIR and ECLASSDIR.
Commit fa4ac9474048ec75af138fc61f22485c06aac5b7 had inadvertently
changed the scope of the PORTDIR and ECLASSDIR variables which were
referring to FILESDIR by a "ditto" in the second column. Restore both
variables to src_*. To this end, reorder variables such that DISTDIR
follows FILESDIR, and move the remark about accessing the directory
into the footnote.

Note: Similar to FILESDIR, accessing DISTDIR in global scope is needed
for assignment of the PATCHES variable in EAPI 6.

Thanks to mgorny for pointing this out.