dependencies.tex: Record slot for := at build time, not install time
Correct the wording for the 'equals' slot operator to indicate that
the slot of the best installed package at the time package is *built*
matters. After all, the common use case for := is to indicate that
the build binds to a specific package version.

This is also more in line with dependency classes where we clearly
split the dependencies between build and runtime, and indicate that only
DEPEND is guaranteed to be available before src_*.