dependencies.tex: Explicitly require matching package for :=
Explicitly indicate that the ebuild is required to provide appropriate
DEPEND for := to have a matching package installed. Additionally ban
the cases where := can not be reliably used or does not make any sense:
PDEPEND (because by intent the dependency can not be in DEPEND),
and any-of groups (because the non-binding nature of any-of group
collides with binding nature of :=).