A is legal in pkg_nofetch.
It is quite common and useful for existing ebuilds to access the A
variable in pkg_nofetch. Also package managers define the variable
there. Update the spec accordingly.

The same applies to AA (which is deprecated, but there is no reason
why its scope should be different from A).

Thanks to Michał Górny for pointing this out.