dependencies: Add extra vert spacing in phase-dep table
Add additional vertical space between rows of the 'Dependency classes
required to be satisfied for a particular phase function' to improve
readbility. For all but the last row, the phase function cell spans
multiple lines while the dependency class cell is one line only. Without
the extra spacing, it is not apparent at a first glance whether
a particular phase function uses dependencies above or below it.