pkg-mgr-commands: Redefine dolib* in terms of libdir.a
Since EAPI 7 removes dolib, it would be a bad fit to describe the two
other commands on top of one that is removed. Instead, start by defining
libdir.a and then define the two other commands on top of it.

This also clarifies that dolib.a & dolib.so do not respect libopts,
while the old wording was not clear on that. However, this is at least
how it was implemented in Portage (with dolib.a and dolib.so overriding
libopts completely).