nonfatal: Reword argument handling to be cleaner
Reword the description of 'nonfatal' to explicitly indicate that it
takes 1+ args and executes them as a command.  The original 'remainder
of its argument' wording sounds weird, given that there is no earlier
mention of argument processing (-> remainder from what?).  This also
clarifies that calling nonfatal with no argument is not valid.