*into, *opts: Clarify the number of arguments
Clarify the number of arguments taken by *into and *opts.  The former
always takes a single argument, so clarify that calling it without
arguments is invalid.  The latter may take 0+ arguments (with 0
resetting the list), so clarify that the correct way is to pass multiple
arguments rather than a single quoted argument.