docompress: Specify the behavior for compressed files more strictly
Specify the behavior for dealing with already compressed files more
strictly, by requiring them not to be compressed twice if compressed
using the same format already.  This replaces the previous vague
statement of 'behaving sensibly if already compressed'.  This statement
had a number of problems.

Firstly, it was entirely vague on what 'sensible behavior' is.  Most
of the developers seemed to presume it means decompressing the file
and compressing it again.  However, without a clearly defined list
of compression formats to support, this is impossible to implement
properly.

Secondly, different package managers disagreed on the implementation.
For example, Portage recompressed the file, with the implementation
originally supporting .Z, .gz and .bz2, and later gaining support for
.lzma and .xz (independently of EAPI support for those formats).
PkgCore and Paludis never supported controllable compression, so they
always installed the files in original format.

The new wording aims to be more precise and strict while preserving
the spirit of the original and allowing the behavior currently exhibited
by the package manager implementations.  Most notably, it accounts for
the possibility of the same file being iterated over by the compression
routine twice.

It does not require the package manager to recognize arbitrary
compressed file formats, making the current implementations compliant
with it.  At the same time, it implies that the developers need to
account for the package manager not handling the particular compressed
file format used upstream, and appropriately needing to decompress
the file in order to fully respect controllable compression.

Signed-off-by: Michał Górny <mgorny@gentoo.org>