dependencies: Permit {,B}DEPEND in source-initiated pkg_setup
Account for build dependencies being satisfied in pkg_setup, when
executed as part of source build, in order to match the realistic
implementation and widespread use in Gentoo.  This changes pkg_setup
from the previous state of no dependencies being guaranteed.

Firstly, this matches the standard package manager implementation
better.  Since pkg_setup is executed as part of the sequence of
successive phases comprising a source build, it is natural that build
dependencies are installed already when it's executed, and not between
pkg_setup and src_unpack.

Secondly, it accounts for widespread use of pkg_setup to determine
installed dependencies in order to set the build environment
appropriately.  This is a case e.g. with python-any-r1 or llvm eclasses,
which query installed packages in order to determine the Python or LLVM
version to use, appropriately.  This behavior can't be changed without
major changes to a large number of ebuilds, and without making ebuild
development harder.

Technically, we could extend the same rule to RDEPEND being satisfied
when package is going to be installed.  However, this would restrict
the flexibility of parallel builds.  With regard to the current rules,
the package manager can start building package from sources before all
of its runtime dependencies are installed.  The unclear benefit
of having RDEPEND satisfied in pkg_setup does not seem to justify losing
this possibility.

Signed-off-by: Michał Górny <mgorny@gentoo.org>
Signed-off-by: Ulrich Müller <ulm@gentoo.org>