doins, dodoc: Clarify how directories are created.
With the -r option, it was unspecified what the mode of any created
directories is. Clarify that doins -r will create them as if dodir was
called (i.e., respect diropts), and that dodoc -r will create them as
if plain install -d was used.

For doins, this agrees with package manager implementations.

For dodoc, this agrees with historic Paludis behaviour. Portage
behaviour has changed in the past, when dodoc was changed from a
standalone helper to reusing parts of doins. Usage in the Gentoo
repository indicates that no ebuilds call diropts specifically for
installing of documentation. However, for several ebuilds dodoc -r is
affected by diropts called previously for another directory, which
looks like an unwanted side effect.

Signed-off-by: Ulrich Müller <ulm@gentoo.org>