ebuild-functions.tex: Update array detection code.
Remove the space after "declare -a" for matching "declare -p" output.
With the update of the bash version to 4.2 in EAPI 6, variables can
have other attributes in addition, which would make the test fail.
For example, see https://bugs.gentoo.org/444832#c7. The implementation
in Portage already omits the space.

Replace grep by functionally equivalent code in bash. This is how the
test is implemented in package managers, and follows the principle
that external programs should not be called unnecessarily.

Redirect stderr for "declare -p", because it outputs an error message
if the PATCHES variable is not defined.

Signed-off-by: Ulrich Müller <ulm@gentoo.org>