pms.cls: Drop the caption package because it interferes with TeX4ht.
The "caption" package already caused issues in the past (for example,
see commits b35f619 and 467f1b7), and version v3.4h finally broke the
list of tables in HTML output. TeX4ht upstream says that the package
should be avoided: https://puszcza.gnu.org.ua/bugs/?313#comment8

Positioning the caption above the table is simple enough without using
the package. So we only lose the boldface labels which is a very minor
issue.

As an added bonus, this allows removal of most workarounds that were
necessary for TeX4ht.

Signed-off-by: Ulrich Müller <ulm@gentoo.org>