pkg-mgr-commands.tex: Clarify "skipped silently" in unpack
The previous wording was added in commit 634c32f with the EAPI 3
(later to become EAPI 4) draft. The commit message doesn't give any
rationale for "silently". It may well be that the wording is a remnant
of the rejected "unpack --if-compressed" feature.

See also the discussion in the 2009-04-23 council meeting:
https://projects.gentoo.org/council/meeting-logs/20090423.txt
(starting at 21:35).

Signed-off-by: Ulrich Müller <ulm@gentoo.org>