tree-layout.tex: Missing eapi is not taken from master repo
The wording on this was not clear: One the one hand we had "if no eapi
file is present, EAPI 0 shall be used", on the other hand "contents of
these files are to be taken from [...] the master repository".

Portage uses EAPI 0 if profiles/eapi is missing in a non-stand-alone
repository.

Signed-off-by: Ulrich Müller <ulm@gentoo.org>